+++ price-feeder/oracle/provider/binance.go
@@ -80,7 +80,11 @@ type (
 	}
 )
 
-func NewBinanceProvider(ctx context.Context, logger zerolog.Logger, pairs ...types.CurrencyPair) (*BinanceProvider, error) {
+func NewBinanceProvider(
+	ctx context.Context,
+	logger zerolog.Logger,
+	pairs ...types.CurrencyPair,
+) (*BinanceProvider, error) {
 	wsURL := url.URL{
 		Scheme: "wss",
 		Host:   binanceHost,
@@ -374,13 +378,13 @@ func (p *BinanceProvider) keepReconnecting() {
 	for time := range reconnectTicker.C {
 		if err := p.reconnect(); err != nil {
 			p.logger.Err(err).Msgf("attempted to reconnect %d times at %s", connectionTries, time.String())
+			connectionTries++
 			continue
 		}
 
 		if connectionTries > maxReconnectionTries {
 			p.logger.Warn().Msgf("failed to reconnect %d times", connectionTries)
 		}
-		connectionTries++
 		return
 	}
 }